requirements.txt
langchain-community==0.0.10
langdetect==1.0.9
#langsmith==0.1.71
mistralai==0.0.11
mistune==3.0.2
ml-dtypes==0.2.0
numpy==1.26.2
py==1.11.0
pylance==0.9.7
pyls-spyder==0.4.0
requests-oauthlib==1.3.1
streamlit==1.29.0
transformers==4.36.2
virtualenv==20.25.0
webencodings==0.5.1
websockets==12.0

